browser check when using 'location'
Build was failing bc 'location' was undefined.  I suspect that the reason `this.props.location` was undefined was bc reach/router which is used by gatsby was not assigning location yet bc it sniffed that it wasn't on a browser environment.  It doesn't make sense for location to be available on a non-browser environment anyways.